+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- `rtic2` feature, renamed `rtic` to `rtic1`
 - `embedded_io::Read` trait implemented for `serial`
 
 ## [v0.11.0] - 2025-09-09